fixing error with removing in undo processes
git-svn-id: svn://107.170.10.188:8080/Semoss/dev@13510 ef391c1c-ee15-45f7-b1d6-dde3d27a2169

Former-commit-id: 965cc77d6b42fcc434f13ea19741acff58a3ef1f

+++ src/prerna/om/Insight.java
@@ -907,7 +907,7 @@ public class Insight {
 			ISEMOSSTransformation transToUndo = trans.get(indexToRemove);
 			transToUndo.setDataMakers(this.dataMaker);
 			transToUndo.undoTransformation();
-			trans.remove(indexToRemove);
+			trans.remove(indexToRemove.intValue());
 			if(transToUndo instanceof JoinTransformation){
 				removedJoin = true;
 			}
